+++ CHANGELOG.md
@@ -1,10 +1,32 @@
 # Change log
 
+## v1.1
+
+### New Features
+
+
+- Recovery of the token with error message in case of failure 
+- Reading options and arguments
+
+### Added:
+
+- File ’parser.py’
+- File ’notion_api.py’
+
 ## v1.0 
 
-Initial release.
+### New Features
+
+- Notion database connexion
+- Notion read and write on the database
+- Reading json file
+- Property formating
 
 ### Added:
 
 - File ’__init__.py’
 - File ’main.py’
+- File ’page.py’
+- File ’database.py’
+- File ’properties.py’
+- File ’request_notion.py’